first pass at rework

+++ src/main/java/bio/terra/service/snapshot/flight/delete/DeleteSnapshotPrimaryDataGcpStep.java
@@ -4,14 +4,10 @@ import bio.terra.common.FlightUtils;
 import bio.terra.common.exception.PdaoException;
 import bio.terra.service.configuration.ConfigEnum;
 import bio.terra.service.configuration.ConfigurationService;
-import bio.terra.service.dataset.Dataset;
-import bio.terra.service.dataset.DatasetService;
 import bio.terra.service.dataset.exception.DatasetNotFoundException;
 import bio.terra.service.filedata.google.firestore.FireStoreDao;
-import bio.terra.service.filedata.google.firestore.FireStoreDependencyDao;
 import bio.terra.service.snapshot.Snapshot;
 import bio.terra.service.snapshot.SnapshotService;
-import bio.terra.service.snapshot.SnapshotSource;
 import bio.terra.service.snapshot.exception.SnapshotNotFoundException;
 import bio.terra.service.tabulardata.google.BigQueryPdao;
 import bio.terra.stairway.FlightContext;
@@ -24,32 +20,26 @@ import java.util.concurrent.TimeUnit;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class DeleteSnapshotPrimaryDataStep implements Step {
+public class DeleteSnapshotPrimaryDataGcpStep implements Step {
 
-  private static Logger logger = LoggerFactory.getLogger(DeleteSnapshotPrimaryDataStep.class);
+  private static Logger logger = LoggerFactory.getLogger(DeleteSnapshotPrimaryDataGcpStep.class);
 
   private BigQueryPdao bigQueryPdao;
   private SnapshotService snapshotService;
-  private FireStoreDependencyDao dependencyDao;
   private FireStoreDao fileDao;
   private UUID snapshotId;
-  private DatasetService datasetService;
   private ConfigurationService configService;
 
-  public DeleteSnapshotPrimaryDataStep(
+  public DeleteSnapshotPrimaryDataGcpStep(
       BigQueryPdao bigQueryPdao,
       SnapshotService snapshotService,
-      FireStoreDependencyDao dependencyDao,
       FireStoreDao fileDao,
       UUID snapshotId,
-      DatasetService datasetService,
       ConfigurationService configService) {
     this.bigQueryPdao = bigQueryPdao;
     this.snapshotService = snapshotService;
-    this.dependencyDao = dependencyDao;
     this.fileDao = fileDao;
     this.snapshotId = snapshotId;
-    this.datasetService = datasetService;
     this.configService = configService;
   }
 
@@ -68,14 +58,13 @@ public class DeleteSnapshotPrimaryDataStep implements Step {
       }
 
       Snapshot snapshot = snapshotService.retrieve(snapshotId);
-      bigQueryPdao.deleteSnapshot(snapshot);
-
-      // Remove snapshot file references from the underlying datasets
-      for (SnapshotSource snapshotSource : snapshot.getSnapshotSources()) {
-        Dataset dataset = datasetService.retrieve(snapshotSource.getDataset().getId());
-        dependencyDao.deleteSnapshotFileDependencies(dataset, snapshotId.toString());
+      if (snapshot.getProjectResource().getGoogleProjectId() != null) {
+        bigQueryPdao.deleteSnapshot(snapshot);
+        fileDao.deleteFilesFromSnapshot(snapshot);
+      } else {
+        logger.info("Google project Id is null, so assume this this an Azure project.");
       }
-      fileDao.deleteFilesFromSnapshot(snapshot);
+
     } catch (BigQueryException ex) {
       if (FlightUtils.isBigQueryIamPropagationError(ex)) {
         return new StepResult(StepStatus.STEP_RESULT_FAILURE_RETRY, ex);
